update test app
