+++ apps/teams-test-app/src/components/AppAPIs.tsx
@@ -1,4 +1,4 @@
-import { app, core, DeepLinkParameters } from '@microsoft/teams-js';
+import { app } from '@microsoft/teams-js';
 import React, { ReactElement } from 'react';
 
 import { ApiWithoutInput, ApiWithTextInput } from './utils';
@@ -15,8 +15,8 @@ const GetContext = (): ReactElement =>
 
 const ExecuteDeepLink = (): ReactElement =>
   ApiWithTextInput<string>({
-    name: 'executeDeepLink2',
-    title: 'Execute Deep Link',
+    name: 'openLink',
+    title: 'Open Link',
     onClick: {
       validateInput: input => {
         if (typeof input !== 'string') {
@@ -24,25 +24,8 @@ const ExecuteDeepLink = (): ReactElement =>
         }
       },
       submit: async input => {
-        await core.executeDeepLink(input);
+        await app.openLink(input);
         return 'Completed';
-      },
-    },
-  });
-
-const ShareDeepLink = (): ReactElement =>
-  ApiWithTextInput<DeepLinkParameters>({
-    name: 'core.shareDeepLink',
-    title: 'core.shareDeepLink',
-    onClick: {
-      validateInput: input => {
-        if (!input.subEntityId || !input.subEntityLabel) {
-          throw new Error('subEntityId and subEntityLabel are required.');
-        }
-      },
-      submit: async input => {
-        await core.shareDeepLink(input);
-        return 'called shareDeepLink';
       },
     },
   });
@@ -62,7 +45,6 @@ const AppAPIs = (): ReactElement => (
     <h1>app</h1>
     <GetContext />
     <ExecuteDeepLink />
-    <ShareDeepLink />
     <RegisterOnThemeChangeHandler />
   </>
 );
